Tests/LibWeb/Text/expected/quirks-mode-case-insensitive-id-selector.txt

ParentNode.querySelector matches ID selectors case-insensitively in quirks mode: true
